[#] Load Levels From Level Select
*At game start a level is automatically loaded and can be played by
switching the view to the map view with "m". Before pressing "m" a
player can now select any level they want from the list and load it for
play by clicking "enter"

*Players can view the map list at any time by pressing "i". They can
then move up and down the list to select a level and load it the same as
desribed above.